# src/main/resources/api/dev-practice.openapi.yaml
openapi: 3.0.3
info:
  title: La Forge API
  version: 0.1.0
  description: |
    Contrat du prototype de laboratoire d’entraînement technique. Texte métier en français, identifiants API en anglais.

    Périmètre : profil, catalogue, exercices versionnés, tentatives, indices, corrigés, révisions, veille, défis collectifs, édition et tuteur IA optionnel. Pas d’exécution distante de code, génération automatique de contenus, classement RH ou gestion multi-entreprises. Une installation correspond à une entreprise ou un groupe pilote. Les rôles sont gérés hors API.

    Authentification : compte créé localement via /auth/registrations (email + mot de passe), activé par confirmation
    du jeton de vérification reçu par courriel (/auth/verifications). Connexion via /auth/sessions : le serveur émet
    lui-même un bearer JWT de courte durée (signature RS256) accompagné d’un jeton de renouvellement opaque
    (/auth/sessions/refresh). Le serveur valide signature, issuer, audience et expiration ; utilisateur courant déduit
    de sub, jamais du corps de requête. Rôles LEARNER, AUTHOR, REVIEWER, ADMIN cumulatifs ; tous les comptes peuvent
    pratiquer dès activation. x-required-roles signifie au moins un rôle de la liste, plus les contraintes de propriété
    décrites. ADMIN n’accorde pas l’accès aux tentatives privées des autres.

...

...

...

...

...

...

...
servers:
- url: /api/v1
  description: Même origine que le frontend ; préfixe de la v1.
security:
- bearerAuth: []
...
paths:
...
  /topics:
    get:
      tags:
      - Catalog
      operationId: listTopics
      summary: Lister les thèmes
      description: Référentiel initial chargé avec le contenu du pilote. Ordre label puis id ; ajout via
        POST /topics.
      responses:
        '200':
          description: Lister les thèmes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
    post:
      tags:
      - Catalog
      operationId: createTopic
      summary: Ajouter un thème
      description: Réservé aux rôles AUTHOR et ADMIN. L’identifiant est généré par le serveur. Un slug déjà
        utilisé produit un conflit.
      responses:
        '201':
          description: Thème créé.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Topic'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TopicInput'
      x-required-roles:
      - AUTHOR
      - ADMIN
  /exercises:
    get:
      tags:
      - Catalog
      operationId: listExercises
      summary: Lister les exercices publiés
      description: Dernière version publiée de chaque exercice. Filtres combinés par ET ; q cherche dans titre et
        énoncé, jamais dans le corrigé. Ordre publishedAt décroissant puis id.
      responses:
        '200':
          description: Lister les exercices publiés.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExercisePage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
      - name: type
        in: query
        schema:
          $ref: '#/components/schemas/ExerciseType'
        description: ''
      - name: difficulty
        in: query
        schema:
          $ref: '#/components/schemas/Difficulty'
        description: ''
      - name: topicId
        in: query
        schema:
          type: string
          format: uuid
        description: ''
      - name: technology
        in: query
        schema:
          type: string
          maxLength: 60
        description: ''
      - name: q
        in: query
        schema:
          type: string
          maxLength: 200
        description: ''
  /exercises/{exerciseId}:
    get:
      tags:
      - Catalog
      operationId: getLatestExercise
      summary: Lire la dernière version publiée
      description: Sans indices ni correction. L’identifiant de version doit être conservé lors de la création d’une
        tentative.
      responses:
        '200':
          description: Lire la dernière version publiée.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exercise'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - name: exerciseId
        in: path
        required: true
        schema:
          type: string
          format: uuid
  /exercises/{exerciseId}/versions/{version}:
    get:
      tags:
      - Catalog
      operationId: getExerciseVersion
      summary: Lire une version publiée précise
      description: Les anciennes versions restent accessibles aux utilisateurs authentifiés. Une publication ultérieure
        ne change jamais cet énoncé.
      responses:
        '200':
          description: Lire une version publiée précise.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exercise'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - name: exerciseId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: version
        in: path
        required: true
        schema:
          type: integer
          format: int32
          minimum: 1
...
  /challenges:
    post:
      tags:
      - Collective
      operationId: createChallenge
      summary: Créer un défi à partager
      description: Tout utilisateur authentifié. Le créateur devient participant. Le code d’invitation n’accorde
        aucun accès anonyme. Le créateur est soumis aux mêmes règles de déblocage que les autres.
      responses:
        '201':
          description: Créer un défi à partager.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengeCreated'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChallengeInput'
  /me/challenges:
    get:
      tags:
      - Collective
      operationId: listMyChallenges
      summary: Lister mes défis
      description: Défis créés ou rejoints, ordre closesAt décroissant puis id. Aucun annuaire des défis privés.
      responses:
        '200':
          description: Lister mes défis.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengePage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
  /challenges/join:
    post:
      tags:
      - Collective
      operationId: joinChallenge
      summary: Rejoindre un défi via son code
      description: 'Code invalide : 404 ; défi fermé : 409. Appel répété pour membre existant : 200, même si fermé.
        Refuse 409 si l’utilisateur a déjà soumis une tentative ou révélé un indice sur cette version hors défi.
        À l’inscription, expliquer que la réponse sera partagée avec les participants ayant soumis.'
      responses:
        '200':
          description: Rejoindre un défi via son code.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Challenge'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JoinChallenge'
  /challenges/{challengeId}:
    get:
      tags:
      - Collective
      operationId: getChallenge
      summary: Lire un défi rejoint
      description: 'Participants uniquement ; autres utilisateurs : 404. responsesUnlocked vrai seulement si l’appelant
        a soumis dans ce défi, même pour le créateur.'
      responses:
        '200':
          description: Lire un défi rejoint.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Challenge'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - name: challengeId
        in: path
        required: true
        schema:
          type: string
          format: uuid
  /challenges/{challengeId}/close:
    post:
      tags:
      - Collective
      operationId: closeChallenge
      summary: Fermer mon défi
      description: 'Créateur uniquement, sinon 403 pour un participant. OPEN → CLOSED ; appel répété : 200. Révoque
        le code, bloque les nouvelles inscriptions et soumissions. Les réponses déjà soumises et discussions restent
        accessibles aux participants ayant soumis.'
      responses:
        '200':
          description: Fermer mon défi.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Challenge'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - name: challengeId
        in: path
        required: true
        schema:
          type: string
          format: uuid
  /challenges/{challengeId}/responses:
    get:
      tags:
      - Collective
      operationId: listSharedResponses
      summary: Comparer les réponses du défi
      description: 'Participant ayant soumis dans ce défi requis, sinon 403 (non membre : 404). Inclut sa propre
        réponse. Ordre submittedAt croissant puis id. Ne révèle aucune autre tentative, statistique personnelle
        ou conversation IA.'
      responses:
        '200':
          description: Comparer les réponses du défi.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SharedResponsePage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - name: challengeId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
  /challenges/{challengeId}/comments:
    get:
      tags:
      - Collective
      operationId: listChallengeComments
      summary: Lire la discussion du défi
      description: Participant ayant soumis requis ; sinon 403, ou 404 si non membre. Ordre createdAt croissant
        puis id.
      responses:
        '200':
          description: Lire la discussion du défi.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - name: challengeId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
    post:
      tags:
      - Collective
      operationId: createChallengeComment
      summary: Participer au débrief collectif
      description: Participant ayant soumis requis ; sinon 403, ou 404 si non membre. Autorisé après fermeture.
        Texte brut, aucun HTML interprété.
      responses:
        '201':
          description: Participer au débrief collectif.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscussionComment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - name: challengeId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentInput'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Jeton du fournisseur d’identité configuré. Acquisition hors périmètre.
  parameters:
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Nombre maximal d’éléments.
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
        maxLength: 1000
      description: Curseur opaque retourné par la page précédente.
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: 'UUID généré par le client, unique pour utilisateur + méthode + chemin. Conservation 24 h. Même
        clé et même corps : réponse originale sans répéter l’effet. Autre corps : 409. Déduplication atomique, y
        compris des appels IA. Opération encore en cours : 409 INVALID_STATE, réessayer avec la même clé.'
  responses:
    BadRequest:
      description: Requête syntaxiquement invalide.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthenticated:
      description: Jeton absent, expiré ou invalide.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
      headers:
        WWW-Authenticate:
          description: Challenge Bearer.
          schema:
            type: string
          example: Bearer
    Forbidden:
      description: Rôle insuffisant ou déblocage pédagogique non acquis.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Ressource inexistante ou non accessible dans le périmètre de l’appelant.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: Transition interdite, tentative déjà finalisée, version de base obsolète ou conflit d’idempotence.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    ValidationFailed:
      description: Données ou contraintes métier invalides ; violations détaille les champs.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    TooManyRequests:
      description: Quota dépassé ; respecter Retry-After.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
      headers:
        Retry-After:
          description: Délai minimal en secondes avant nouvel essai, si disponible.
          schema:
            type: integer
            format: int32
            minimum: 0
    InternalError:
      description: Erreur serveur inattendue.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
...
    ExerciseType:
      type: string
      enum:
      - CODE_REVIEW
      - PREDICTION
      - DIAGNOSIS
      - REFACTORING
      - TECH_DISCOVERY
      - TECHNICAL_CHOICE
      - QUIZ
    Difficulty:
      type: string
      enum:
      - BEGINNER
      - INTERMEDIATE
      - ADVANCED
...
    Topic:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
          maxLength: 80
          pattern: ^[a-z0-9-]+$
        label:
          type: string
          maxLength: 200
          minLength: 1
      required:
      - id
      - slug
      - label
    TopicInput:
      type: object
      additionalProperties: false
      properties:
        slug:
          type: string
          maxLength: 80
          pattern: ^[a-z0-9-]+$
        label:
          type: string
          maxLength: 200
          minLength: 1
      required:
      - slug
      - label
...
    Exercise:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        version:
          type: integer
          format: int32
          minimum: 1
        title:
          type: string
          maxLength: 200
          minLength: 1
        type:
          $ref: '#/components/schemas/ExerciseType'
        difficulty:
          $ref: '#/components/schemas/Difficulty'
        topicIds:
          type: array
          items:
            type: string
            format: uuid
          maxItems: 20
          uniqueItems: true
        estimatedMinutes:
          type: integer
          format: int32
          minimum: 1
          maximum: 60
        technologies:
          type: array
          items:
            $ref: '#/components/schemas/TechnologyRequirement'
          maxItems: 10
        publishedAt:
          type: string
          format: date-time
        promptMarkdown:
          type: string
          maxLength: 20000
          minLength: 1
        learningObjectives:
          type: array
          items:
            type: string
            maxLength: 200
            minLength: 1
          maxItems: 10
          minItems: 1
        files:
          type: array
          items:
            $ref: '#/components/schemas/CodeFile'
          maxItems: 10
        responseSpec:
          $ref: '#/components/schemas/ResponseSpec'
        hintCount:
          type: integer
          format: int32
          minimum: 0
          maximum: 5
      required:
      - id
      - version
      - title
      - type
      - difficulty
      - topicIds
      - estimatedMinutes
      - technologies
      - publishedAt
      - promptMarkdown
      - learningObjectives
      - files
      - responseSpec
      - hintCount
      description: Projection apprenant. Ne contient ni corrigé, ni bonne réponse, ni contenu des indices non débloqués.
        Une version publiée est immuable.
...
    ChallengeInput:
      type: object
      additionalProperties: false
      properties:
        title:
          type: string
          maxLength: 200
          minLength: 1
        exerciseId:
          type: string
          format: uuid
        exerciseVersion:
          type: integer
          format: int32
          minimum: 1
        closesAt:
          type: string
          format: date-time
      required:
      - title
      - exerciseId
      - exerciseVersion
      - closesAt
      description: Échéance strictement future, au maximum dans 30 jours. Défi ouvert immédiatement. Version publiée
        uniquement. Créateur inscrit automatiquement ; titre/version/échéance immuables.
    Challenge:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          maxLength: 200
          minLength: 1
        exercise:
          $ref: '#/components/schemas/ExerciseSummary'
        creatorId:
          type: string
          format: uuid
        closesAt:
          type: string
          format: date-time
        state:
          type: string
          enum:
          - OPEN
          - CLOSED
        participantCount:
          type: integer
          format: int32
          minimum: 1
        responsesUnlocked:
          type: boolean
      required:
      - id
      - title
      - exercise
      - creatorId
      - closesAt
      - state
      - participantCount
      - responsesUnlocked
    ChallengeCreated:
      type: object
      additionalProperties: false
      properties:
        challenge:
          $ref: '#/components/schemas/Challenge'
        joinCode:
          type: string
          maxLength: 128
          minLength: 32
      required:
      - challenge
      - joinCode
      description: Code opaque aléatoire, au moins 128 bits d’entropie. Transmettre dans le fragment du lien frontend
        puis dans le corps de join. Ne pas journaliser. Rendu uniquement à la création, valable jusqu’à closesAt.
        Le créateur peut fermer le défi pour le révoquer.
    JoinChallenge:
      type: object
      additionalProperties: false
      properties:
        joinCode:
          type: string
          maxLength: 128
          minLength: 32
      required:
      - joinCode
...
    CommentInput:
      type: object
      additionalProperties: false
      properties:
        body:
          type: string
          maxLength: 4000
          minLength: 1
      required:
      - body
...
    TopicPage:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Topic'
          maxItems: 100
        nextCursor:
          type: string
          maxLength: 1000
      required:
      - items
      description: Pagination par curseur opaque ; nextCursor absent à la dernière page. Aucun total garanti. Un
        curseur doit être réutilisé avec les mêmes filtres.
    ExercisePage:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ExerciseSummary'
          maxItems: 100
        nextCursor:
          type: string
          maxLength: 1000
      required:
      - items
      description: Pagination par curseur opaque ; nextCursor absent à la dernière page. Aucun total garanti. Un
        curseur doit être réutilisé avec les mêmes filtres.
...
    ChallengePage:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Challenge'
          maxItems: 100
        nextCursor:
          type: string
          maxLength: 1000
      required:
      - items
      description: Pagination par curseur opaque ; nextCursor absent à la dernière page. Aucun total garanti. Un
        curseur doit être réutilisé avec les mêmes filtres.
    SharedResponsePage:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SharedResponse'
          maxItems: 100
        nextCursor:
          type: string
          maxLength: 1000
      required:
      - items
      description: Pagination par curseur opaque ; nextCursor absent à la dernière page. Aucun total garanti. Un
        curseur doit être réutilisé avec les mêmes filtres.
    CommentPage:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/DiscussionComment'
          maxItems: 100
        nextCursor:
          type: string
          maxLength: 1000
      required:
      - items
      description: Pagination par curseur opaque ; nextCursor absent à la dernière page. Aucun total garanti. Un
        curseur doit être réutilisé avec les mêmes filtres.
...